requirements.txt
arxiv
pymupdf
sentence-transformers
chromadb
google-generativeai
gradio
python-dotenv
langchain